Tele/for_computer/requirements.txt
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
pyserial==3.5
ultralytics==8.0.136
